[BUZZOK-31177] Rename _kickoff_storage module to kickoff_storage
Review feedback: a public module shouldn't start with an underscore.
